+++ .gitignore
@@ -1,5 +1,4 @@
 /.idea/
 /cmake-build-debug/
 /cmake-build-release/
-
 *.exe
