+++ ClassDesign/DataHelper/DataType.cpp
@@ -1,0 +1,1 @@
+#include "DataType.h"
