--- ClassDesign/DataType.cpp
@@ -1,2 +1,0 @@
-#include "DataType.h"
-#include "Player/ChessTreeRobot.h"
